Test Huds
look at the Hud 1 and 2 scenes, in the game window. Be sure to lock the
aspect ratio to 16:9. We can tool around with either way to make it look
good, we just have to decide on a style of level design. Also, git
ignore update didn't work :/ no idea how to ignore the Library files.