holy crap lots of changes
need to start the game from the menu, and need to have 2 players to
start a game, or it would end immediately. Still haven't locked a stage
design, we need to discuss more. Also, gameplay will not be close to
finished for a good while, still a ton to add and constant tweaking.